# src/main/resources/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Esta é a documentação swagger da API para o TCC e-commerce de chinelos, ainda não está implementada e ainda faltam algumas actions, mas pretendemos continuar o projeto até o fim.",
    "version": "1.0.0",
    "title": "Slippers Swagger",
    "contact": {
      "email": "andrey_coradin@hotmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api",
...
  "paths": {
...
    "/tamanho/listaTamanhos": {
      "get": {
        "tags": [
          "Tamanho"
        ],
        "summary": "Busca de Tamanhos",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Tamanho"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          }
        }
      }
    },
    "/tamanho/inserirTamanho": {
      "post": {
        "tags": [
          "Tamanho"
        ],
        "summary": "Inserção de tamanho",
        "description": "Apenas ADM tem acesso à esta função",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para criação de tamanho",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TamanhoForm"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Tamanho"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/tamanho/alterarTamanho/{id}": {
      "put": {
        "tags": [
          "Tamanho"
        ],
        "summary": "Atualização de tamanhos",
        "description": "Apenas ADM tem acesso à esta função",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "id do tamanho a ser atualizado",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Objeto para alteração de tamanho",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TamanhoForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Tamanho"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/400"
            }
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/tamanho/deletarTamanho/{id}": {
      "delete": {
        "tags": [
          "Tamanho"
        ],
        "summary": "Deleção de tamanho",
        "description": "Apenas ADM tem acesso à esta função",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "description": "id do tamanho a ser deletado",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Apagado com sucesso"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Deves colocar o token do usuário logado na header da requisição HTTP"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira aqui o token resgatado.\nFormato: Bearer {token}\nExemplo: Bearer n8we9f9iom1f3289fj12f2mof123f9m2.f23"
    }
  },
  "definitions": {
    "400": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Mensagem de erro"
        }
      }
    },
...
    "Tamanho": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "descricao": {
          "type": "string",
          "example": "40"
        }
      }
    },
...
    "TamanhoForm": {
      "type": "object",
      "properties": {
        "descricao": {
          "type": "string",
          "example": "39/40"
        }
      }
    },
...
  }
}